Improve Code Quality 5 (#259)
* [Tests] Minor formats

* Eliminate unnecessary X509Certificate2 allocations

* Move properties below constructor

* Presize CborArray when size is known

* Format CredentialPublicKey

* Use init properties in AuthenticatorAssertionResponse

* Format AttestationFormat

* Use DateTimeOffset.UnixEpoch and differentiate between missing and invalid timestampMs

* Remove unused using statement

* Remove leftover test logic

+++ Test/Attestation/Apple.cs
@@ -60,11 +60,11 @@ namespace Test.Attestation
 
                     var cpk = new CborMap
                     {
-                        { (int)COSE.KeyCommonParameter.KeyType, (int)type },
-                        { (int)COSE.KeyCommonParameter.Alg, (int)alg },
-                        { (int)COSE.KeyTypeParameter.X, ecparams.Q.X },
-                        { (int)COSE.KeyTypeParameter.Y, ecparams.Q.Y },
-                        { (int)COSE.KeyTypeParameter.Crv, (int)crv }
+                        { COSE.KeyCommonParameter.KeyType, type },
+                        { COSE.KeyCommonParameter.Alg, alg },
+                        { COSE.KeyTypeParameter.X, ecparams.Q.X },
+                        { COSE.KeyTypeParameter.Y, ecparams.Q.Y },
+                        { COSE.KeyTypeParameter.Crv, crv }
                     };
 
                     var x = (byte[])cpk[COSE.KeyTypeParameter.X];
